pruebas unitarias y dtos

+++ src/main/java/com/backed/main/controller/categoriaController.java
@@ -34,7 +34,7 @@ public class categoriaController {
 
     @ApiOperation(value = "Se obtiene una categoria por id")
     @GetMapping( path = "/{id}" )
-    public java.util.Optional<categoriaEntity> obtenerCategoriaPorId(@PathVariable("id") Long id) {
+    public java.util.Optional<categoriaEntity> obtenerCategoriaPorId(@PathVariable("id") int id) {
         return this.categoraService.obtenerCategoriaPorId(id);
     }
